Moved command client
