+++ apps/visualstudio/visual_studio.talon
@@ -5,6 +5,7 @@ tag(): user.tabs
 tag(): user.line_commands
 tag(): user.find_and_replace
 tag(): user.multiple_cursors
+tag(): user.command_client
 #multiple_cursor.py support end
 
 # Panels
