+++ core/command_client/rpc_client/get_communication_dir_path.py
